Add files via upload
Major change: some of the functionalities of class ReadProcessByChunks are now placed in two auxiliary classes: class Log and class Reader. The former focuses on maintaining the log file solely; the latter takes care of the various situations how the input files are read into the memory as chunks. Thus, the primary class ReadProcessByChunks does not need to be concerned about file input. It is majorly responsible for receiving the chunks from the Reader object, distributing the work load, and assembling the final results